feat: enhance order management with CRUD operations, improved validation, and error handling

+++ src/modules/order/order.validation.ts
@@ -5,7 +5,7 @@ export const createOrderSchema = z.object({
   products: z.array(
     z.object({
       productId: z.string().min(1, 'Product ID is required'),
-      quantity: z.number().int().min(1, 'Quantity must be at least 1'),
+      quantity: z.number().min(1, 'Quantity must be at least 1'), // assuming quantity is a string in your schema
     })
   ),
   totalPrice: z.number().min(0, 'Total price must be greater than 0'),
